requirements.txt
appnope==0.1.0
astroid==2.4.2
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
faiss-cpu==1.6.3
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
jupyter-client==6.1.6
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.5
numpy==1.19.1
packaging==20.4
pandas
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.2
sentence-transformers==0.3.3
sentencepiece==0.1.91
six==1.15.0
streamlit
threadpoolctl==2.1.0
tokenizers==0.8.1rc1
toml==0.10.1
torch==1.6.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.0.2
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1